% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kaka.get_config}
\alias{kaka.get_config}
\title{Getting configuration for an experiment and data_source}
\usage{
kaka.get_config(realm, experiment, data_source, host = kaka.config.host,
  port = kaka.config.port)
}
\arguments{
\item{realm}{A data realm}

\item{experiment}{A name of an experiment}

\item{data_source}{A name of a data source}

\item{host}{Web host to talk to (default web)}

\item{port}{Port to web host (default 80, but often 8001)}
}
\value{
data_frame
}
\description{
Getting configuration for an experiment and data_source
}
\examples{
dat <- kaka.get_config('genotype', experiment='Gene Expression', data_source='')
}

