% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kaka.qry_json}
\alias{kaka.qry_json}
\title{Querying Kaka using json (helper for kaka.qry)}
\usage{
kaka.qry_json(realm, qry, host = kaka.config.host, port = kaka.config.port)
}
\arguments{
\item{realm}{A data realm}

\item{qry}{A pql query}

\item{host}{Web host to talk to (default web)}

\item{port}{Port to web host (default 80, but often 8001)}
}
\value{
data_frame
}
\description{
Querying Kaka using json (helper for kaka.qry)
}
\examples{
pkgBio('deseq2')
}

