% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kaka.send}
\alias{kaka.send}
\title{Sending data to Kaka}
\usage{
kaka.send(data, config, host = kaka.config.host, port = kaka.config.port)
}
\arguments{
\item{data}{A data frame}

\item{config}{A confiuguration dict (see examples)}

\item{host}{Web host to talk to (default web)}

\item{port}{Port to web host (default 80, but often 8001)}
}
\value{
data_frame
}
\description{
Sending data to Kaka
}
\examples{
dat <- kaka.get_config('genotype', experiment=='Gene Expression', data_source='')
}

