% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pkgBio}
\alias{pkgBio}
\title{Automatically loading/installing biocinductor packages if necessary}
\usage{
pkgBio(x)
}
\arguments{
\item{A}{package name}
}
\value{
Nothing
}
\description{
Automatically loading/installing biocinductor packages if necessary
}
\examples{
pkgBio('deseq2')
}

